stock-trading-analyzer/requirements.txt
yfinance
APScheduler
scikit-learn
XGBoost
requests
pandas
numpy
matplotlib
python-telegram-bot
concurrent.futures